started implementing JPA persistence for nominees & categories

+++ aanestys-backend/src/main/java/org/x_rust/aanestys/samples/backend/DataService.java
@@ -17,9 +17,9 @@ public abstract class DataService {
 
     public abstract void updateProduct(Nominee p);
 
-    public abstract void deleteProduct(int productId);
+    public abstract void deleteProduct(long productId);
 
-    public abstract Nominee getProductById(int productId);
+    public abstract Nominee getProductById(long productId);
 
     public static DataService get() {
         return MockDataService.getInstance();
